backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
requests
alpaca-py
python-dotenv
httpx
websockets
pytz
